move shared methods to utils
